# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MoneyApp API
  version: 1.0.0
  description: |
    Self-hosted budget and net-worth tracker.

    Conventions:
      * Session cookie auth. `user_id` is taken only from the session, never from input.
      * Money is always an object with integer minor units. No floats on the wire.
      * `null` means "not recorded". `0` means "recorded zero". Never interchangeable.
      * Errors are RFC 7807 application/problem+json.
      * Lists are cursor-paginated on (occurred_on, id).
  license:
    name: MIT
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /snapshots:
    get:
      tags: [snapshots]
      summary: Snapshots for a period
      description: Returns every account with the previous month pre-filled, for the entry screen.
      parameters:
        - $ref: "#/components/parameters/PeriodQuery"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Snapshot" }

    post:
      tags: [snapshots]
      summary: Record a whole month at once
      description: |
        The confirm-or-adjust screen, written in one transaction so a half-saved
        month cannot exist. This is the endpoint behind the roughly thirty
        workbook cells the capital block used to need every month.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [period, items]
              properties:
                period: { type: string, pattern: '^\d{4}-\d{2}$' }
                items:
                  type: array
                  items:
                    type: object
                    required: [account_id]
                    properties:
                      account_id:    { type: integer }
                      amount:        { $ref: "#/components/schemas/Money" }
                      quantity_nano: { type: integer, nullable: true }
                      note:          { type: string }
      responses:
        "200":
          description: Written
          content:
            application/json:
              schema:
                type: object
                properties:
                  period:       { type: string }
                  rows_written: { type: integer }
        "422": { $ref: "#/components/responses/Problem" }
  /snapshots/{account_id}/{period}:
    parameters:
      - name: account_id
        in: path
        required: true
        schema: { type: integer }
      - name: period
        in: path
        required: true
        schema: { type: string, pattern: '^\d{4}-\d{2}$' }
    put:
      tags: [snapshots]
      summary: Record one account's balance
      description: |
        Writing to a computed parent is refused with 422: its value is the sum of
        its children. Exactly one of `amount` and `quantity_nano` must be given.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:        { $ref: "#/components/schemas/Money" }
                quantity_nano: { type: integer, nullable: true }
                note:          { type: string }
      responses:
        "200":
          description: Recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Snapshot" }
        "422": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [snapshots]
      summary: Remove one snapshot
      description: A month with no snapshot is a gap. It is never interpolated.
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Problem" }
  /snapshots/{period}/reconciliation:
    parameters:
      - name: period
        in: path
        required: true
        schema: { type: string, pattern: '^\d{4}-\d{2}$' }
    get:
      tags: [snapshots]
      summary: Snapshot versus transaction-implied balance
      description: |
        The snapshot is authoritative. The difference is reported so it can be
        investigated, and nothing is ever corrected automatically
        (docs/adr/0003-snapshot-reconcile-balances.md).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  period: { type: string }
                  note:   { type: string }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        account_id: { type: integer }
                        name:       { type: string }
                        snapshot:
                          oneOf:
                            - $ref: "#/components/schemas/Money"
                            - type: "null"
                        implied:
                          oneOf:
                            - $ref: "#/components/schemas/Money"
                            - type: "null"
                        difference:
                          oneOf:
                            - $ref: "#/components/schemas/Money"
                            - type: "null"
                          description: snapshot minus implied; null when either side is absent.
  /reports/net-worth-in/{currency}:
    parameters:
      - name: currency
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [reports]
      summary: Net worth expressed in another currency
      description: The workbook's rows 71-72, at each period's own rate rather than today's.
      parameters:
        - $ref: "#/components/parameters/PeriodQuery"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  period:        { type: string }
                  currency:      { type: string }
                  base_currency: { type: string }
                  general:
                    oneOf:
                      - $ref: "#/components/schemas/Money"
                      - type: "null"
                  value:
                    oneOf:
                      - $ref: "#/components/schemas/Money"
                      - type: "null"
        "422": { $ref: "#/components/responses/Problem" }
...
  /reports/capital:
    get:
      tags: [reports]
      summary: Net worth, allocation and runway
      parameters:
        - $ref: "#/components/parameters/PeriodQuery"
        - $ref: "#/components/parameters/Valuation"
        - name: burn_mode
          in: query
          schema:
            type: string
            enum: [essential_planned, actual_trailing_3, actual_trailing_12, legacy_blend]
            default: actual_trailing_3
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CapitalReport" }
  /reports/capital/series:
    get:
      tags: [reports]
      summary: Net worth over time, with the real and FX split
      parameters:
        - $ref: "#/components/parameters/FromPeriod"
        - $ref: "#/components/parameters/ToPeriod"
        - name: burn_mode
          in: query
          schema:
            type: string
            enum: [essential_planned, actual_trailing_3, actual_trailing_12, legacy_blend]
            default: actual_trailing_3
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  from:          { type: string }
                  to:            { type: string }
                  base_currency: { type: string }
                  burn_mode:     { type: string }
                  points:
                    type: array
                    items:
                      type: object
                      properties:
                        period:   { type: string }
                        recorded: { type: boolean }
                        general:
                          oneOf:
                            - $ref: "#/components/schemas/Money"
                            - type: "null"
                        ready_for_usage:
                          oneOf:
                            - $ref: "#/components/schemas/Money"
                            - type: "null"
                        runway_months: { type: number, nullable: true }
                        change: { $ref: "#/components/schemas/CapitalChange" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: moneyapp_session
  parameters:
    FromPeriod:
      name: from
      in: query
      schema: { type: string, pattern: "^[0-9]{4}-[0-9]{2}$" }
      description: First month, inclusive. Defaults with `to` to the latest fiscal year.
    ToPeriod:
      name: to
      in: query
      schema: { type: string, pattern: "^[0-9]{4}-[0-9]{2}$" }
      description: Last month, inclusive.
...
    PeriodQuery:
      name: period
      in: query
      required: true
      schema: { type: string, pattern: '^\d{4}-\d{2}$' }
    Valuation:
      name: valuation
      in: query
      description: |
        contemporaneous values each period at its own rate.
        constant values everything at today's rate, reproducing the legacy spreadsheet.
      schema:
        type: string
        enum: [contemporaneous, constant]
        default: contemporaneous
...
  responses:
    Problem:
      description: RFC 7807 problem document
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:   { type: string }
        title:  { type: string }
        status: { type: integer }
        detail: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field:   { type: string }
              message: { type: string }
    Money:
      type: object
      required: [amount_minor, currency, exponent]
      properties:
        amount_minor:
          type: integer
          format: int64
          description: Value in minor units. HUF has exponent 0, so 1000 means 1000 HUF.
        currency: { type: string, minLength: 3, maxLength: 3 }
        exponent: { type: integer, minimum: 0, maximum: 4 }
...
    Snapshot:
      type: object
      properties:
        account_id:      { type: integer }
        account_name:    { type: string }
        period:          { type: string }
        amount:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
        quantity_nano:
          type: integer
          nullable: true
          description: Alternative to amount, for metals and crypto priced by a provider.
        base_amount:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
        previous_amount:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
          description: Prior month, for pre-filling the entry screen.
        asset_class:  { type: string }
        currency:     { type: string }
        is_liquid:    { type: boolean }
        counts_toward_net_worth: { type: boolean }
        computed:
          type: boolean
          description: |
            A parent account. Its value is the sum of its children and a direct
            write is refused with 422.
        value:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
          description: Base-currency value, rolled up for a parent.
        note:         { type: string, nullable: true }
...
    CapitalChange:
      type: object
      description: |
        The month's movement, split into what was actually saved and what the
        currencies did. The workbook reported only their sum, so a month when the
        hryvnia moved looked identical to a month of genuine saving (deviation D8).
      properties:
        recorded:
          type: boolean
          description: false when there is no previous month to compare against.
        total:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
        real:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
        fx:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
    CapitalReport:
      type: object
      properties:
        period:        { type: string }
        base_currency: { type: string }
        general:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
          description: Null for a month with no balances recorded at all.
        ready_for_usage:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
        runway_months:
          type: number
          nullable: true
          description: Null when the burn rate is zero — "forever" is not a number of months.
        burn_mode: { type: string }
        burn_rate:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
        change: { $ref: "#/components/schemas/CapitalChange" }
        allocation:
          type: array
          description: |
            Leaf accounts only, converted to base currency first, so the shares sum
            to exactly 1. The sheet's doughnut plotted parents alongside their
            children and divided unconverted amounts, reaching about 162% with
            `Cash HUF` at 41.4% (deviation D6).
          items:
            type: object
            properties:
              account_id:  { type: integer }
              name:        { type: string }
              asset_class: { type: string }
              value:       { $ref: "#/components/schemas/Money" }
              share:       { type: number }
        accounts:
          type: array
          items: { $ref: "#/components/schemas/Snapshot" }
...